+++ app/src/main/java/com/example/firebase_config/views/MainHomePageActivity.kt
@@ -3,7 +3,6 @@ package com.example.firebase_config.views
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.example.firebase_config.CreatePostActivity
 import com.example.firebase_config.databinding.ActivityMainHomePageBinding
 class MainHomePageActivity : AppCompatActivity() {
