+++ app/src/main/java/com/example/firebase_config/views/authFragments/SignInFragment.kt
@@ -1,11 +1,10 @@
-package com.example.firebase_config.views.fragments
+package com.example.firebase_config.views.authFragments
 
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
 import com.example.firebase_config.databinding.FragmentSignInBinding
