+++ app/src/main/java/com/example/firebase_config/views/createPostFragments/SecondQuestionSurveyFragment.kt
@@ -1,14 +1,11 @@
-package com.example.firebase_config.fragments
+package com.example.firebase_config.views.createPostFragments
 
-import android.os.Build
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.annotation.RequiresApi
 import androidx.fragment.app.Fragment
-import com.example.firebase_config.CreatePostActivity
+import com.example.firebase_config.views.CreatePostActivity
 import com.example.firebase_config.databinding.SecondQuestionSurveyFragmentBinding
 
 class SecondQuestionSurveyFragment : Fragment() {
